chore: move from scss to css; use postcss-nested for nesting
- removes sass from package.json
- uses CSS variables (imported separately in main.jsx)
