+++ src/renderer/src/components/TodoDate.module.css
@@ -1,16 +1,14 @@
-@import "../styles/variables";
-
 .date {
   display: flex;
   align-items: center;
   justify-content: space-between;
   padding: 0 0 40px 0;
-  font-size: $heading-font-size;
+  font-size: var(--heading-font-size);
   .calendar {
     display: flex;
     align-items: center;
     .day {
-      font-size: $big-font-size;
+      font-size: var(--big-font-size);
       font-weight: bold;
     }
     .my {
