+++ src/renderer/src/styles/variables.css
@@ -1,0 +1,30 @@
+:root {
+  --bg: #403f4d;
+  --green: #62dca5;
+  --yellow: #f7f879;
+  --red: #e1675a;
+
+  --input-color: #2e2d33;
+  --item-color: #4e4d5c;
+  --font-color: #fbfafb;
+
+  --big-font-size: 64px;
+  --heading-font-size: 24px;
+  --item-font-size: 20px;
+}
+
+@keyframes click {
+  0% {
+    opacity: 0;
+    width: 0;
+    height: 0;
+  }
+  50% {
+    opacity: 0.5;
+  }
+  100% {
+    opacity: 0;
+    width: 30px;
+    height: 30px;
+  }
+}
